pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fixray"
version = "0.1.0"
description = "AI-powered fracture detection system for X-ray images"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "pillow",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools.packages.find]
where = ["src"]
